requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.24.3
python-multipart==0.0.6
ffmpeg-python==0.2.0
gTTS==2.3.2
pydub==0.25.1
moviepy==1.0.3